Work done 02/04/19
Decided to shelve and not show addative mapping yet

+++ Code/src/LReturn.java
@@ -18,18 +18,18 @@ import java.util.ArrayList;
  */
 
 public class LReturn {
-	private Integer startAngle;
-	private double startScore;
-	private Integer endAngle;
-	private double endScore;
-	private ArrayList<Integer> returnList;
+	private Double startAngle;
+	private Double startScore;
+	private Double endAngle;
+	private Double endScore;
+	private ArrayList<Double> returnList;
 
 	/**
 	 *The blank constructor for an LReturn. Can be used in conjunction with the setter methods
 	 *to set incoming values from the VLSensor and Processor	
 	 */
 	public LReturn () {
-		returnList = new ArrayList<Integer>();
+		returnList = new ArrayList<Double>();
 	}
 
 	/**
@@ -38,9 +38,9 @@ public class LReturn {
 	 * @param start - The starting angle of the blockage
 	 * @param blockage - The distance the start of the blockage is away
 	 */
-	public LReturn (Integer start, Integer blockage) {
+	public LReturn (Double start, Double blockage) {
 		startAngle = start;
-		returnList = new ArrayList<Integer>();
+		returnList = new ArrayList<Double>();
 	}
 
 	/**
@@ -50,7 +50,7 @@ public class LReturn {
 	 * @param blocks - The ArrayList of Blockage distances
 	 * @param end - The ending angle of the blockage
 	 */
-	public LReturn (Integer start, ArrayList<Integer> blocks, Integer end) {
+	public LReturn (Double start, ArrayList<Double> blocks, Double end) {
 		startAngle = start;
 		returnList = blocks;
 		endAngle = end;
@@ -60,7 +60,7 @@ public class LReturn {
 	 * Gets the starting angle
 	 * @return the starting angle
 	 */
-	public Integer getStart() {
+	public Double getStart() {
 		return startAngle;
 	}
 
@@ -69,7 +69,7 @@ public class LReturn {
 	 * @param start - the starting angle of a blockage
 	 * @return true
 	 */
-	public void setStart(Integer start) {
+	public void setStart(Double start) {
 		startAngle = start;
 	}
 
@@ -77,7 +77,7 @@ public class LReturn {
 	 * Gets the ending angle
 	 * @return the ending angle
 	 */
-	public Integer getEnd() {
+	public Double getEnd() {
 		return endAngle;
 	}
 
@@ -86,7 +86,7 @@ public class LReturn {
 	 * @param end
 	 * @return true
 	 */
-	public void setEnd(Integer end) {
+	public void setEnd(Double end) {
 		endAngle = end;
 	}
 
@@ -95,8 +95,8 @@ public class LReturn {
 	 * @param theAngle - the selected angle
 	 * @return the distance of the blockage at the given angle
 	 */
-	public Integer getDistance(Integer theAngle) {
-		return returnList.get(theAngle);
+	public Double getDistance(Double theAngle) {
+		return returnList.get(theAngle.intValue());
 	}
 
 	/**
@@ -104,7 +104,7 @@ public class LReturn {
 	 * @param theDistance - new distance in mm
 	 * @return true
 	 */
-	public void setDistance(Integer theDistance) {
+	public void setDistance(Double theDistance) {
 		returnList.add(theDistance);
 	}
 
@@ -112,7 +112,7 @@ public class LReturn {
 	 * Returns the calculated pathfinding score for the starting angle
 	 * @return the score
 	 */
-	public double getStartScore() {
+	public Double getStartScore() {
 		return startScore;
 	}
 	
@@ -122,7 +122,7 @@ public class LReturn {
 	 * @param score - the calculated score
 	 * @return true
 	 */
-	public void setStartScore(double score) {
+	public void setStartScore(Double score) {
 		startScore = score;
 	}
 	
@@ -130,7 +130,7 @@ public class LReturn {
 	 * Gets the calculated pathfinding score for the end of the blockage
 	 * @return the score
 	 */
-	public double getEndScore() {
+	public Double getEndScore() {
 		return endScore;
 	}
 	
@@ -140,7 +140,7 @@ public class LReturn {
 	 * @param score - the calculated score
 	 * @return true
 	 */
-	public void setEndScore(double score) {
+	public void setEndScore(Double score) {
 		endScore = score;
 	}
 	
@@ -149,7 +149,7 @@ public class LReturn {
 	 * this cleans up the code to get the first item as it will be more commonly used)
 	 * @return the distance to the blockage at the start angle
 	 */
-	public Integer getStartDist() {
+	public Double getStartDist() {
 		return returnList.get(0);
 	}
 
@@ -158,19 +158,19 @@ public class LReturn {
 	 * this cleans up the code to get the last item as it will be more commonly used)
 	 * @return the distance to the blockage at the end angle
 	 */
-	public Integer getEndDist() {
+	public Double getEndDist() {
 		return returnList.get(returnList.size() -1);
 	}
 	
-	public void appendBlocks(ArrayList<Integer> blocks){
+	public void appendBlocks(ArrayList<Double> blocks){
 		returnList.addAll(blocks);
 	}
 	
-	public ArrayList<Integer> getBlocks(){
+	public ArrayList<Double> getBlocks(){
 		return returnList;
 	}
 
 	public void insertZero() {
-		returnList.add(0, 0);
+		returnList.add(0, 0.0);
 	}
 }
